requirements.txt
APScheduler==3.6.3
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
convert-numbers==0.4
dynaconf==3.1.5
idna==3.3
isort==5.10.1
jdatetime==3.6.4
peewee==3.14.4
pendulum==2.1.2
prettytable==2.2.1
pycodestyle==2.8.0
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.3
pytzdata==2020.1
requests==2.26.0
six==1.16.0
toml==0.10.2
tornado==6.1
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.7
wcwidth==0.2.5
